Added a (very) rough way to send Objects between two people.
Right now, any number of pairs of clients can connect to the server and send String objects. All the server does is connect two clients with one another; a Client starting a new 'game' (i.e., chat session) sends -1 as the first bit of data, and a Client joining another Client sends the ID of the 'game' (i.e., chat) he wishes to join. After that, for both Clients, the Server simply reads any Objects sent and sends them to the other Client, so in the future, all that has to be done is, for example, create a 'Message' class that could contain either the a Move or a chat message; the Server doesn't care about types, so it doesn't matter what Objects the Clients are sending.