Made a variety of changes to networking.
Changed the protocol between Clients and the server: if a new game is being created, the Client waits for the server to write back to it the gameID it was given. Also changed the way that the ClientHandler holds waiting clients to fix concurrency issues and allow for an arbitrary number of waiting Clients. (Now, when the clients arrive, they are given a gameID off of a queue of IDs that is established when the Server starts, and when a connection has been established between peers, the gameID is replaced on the queue.)