Made a number of changes to the networking.
The comments for the send methods in Client.java and in the run method in ClientReader.java sort of try to explain how I thought this might work. Initially, something will set up the connection and establish the new game or join the existing game, as can be seen in the MainTest.java file's main method. Then, one thread will always be reading, and when it receives a message, it will hand it to the UI to process, all in that thread (though perhaps that should be changed). Another thread will have a queue of messages that it will take off, as they're available, and send.